base_code/requirements.txt
torch==2.0.0
torchvision==0.15.0
scikit-learn==1.2.2
opencv-python==4.7.0
Pillow==9.5.0